Emit summary reports when summary.rb is called
The interesting report is pr-summary.csv because it ostensibly tells
what wrong things were caused by this PR, such that we cannot merge it
without fixing them.

May also be interested in baseline-unresolved.csv so we can start to
tackle some of the issues that are in the deployed website.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>